Make Debian init files compatible with systemd
In order for systemd to use init scripts in sysv compatibility mode,
specific data is required at the top of the init script.  This provides a
quick workaround for the lack of systemd support for Debian/Ubuntu in this
module.

TODO: Add a proper systemd unit file for Debian family OSes.